initial commit
cash-register-app This is the cash register app.  The program is built using Vanilla JS, HTML and CSS.  The site asks for a bill amount and the cash recieved from the coustomer and gives the currency notes cardinality in a currency table.  I build this project as a part of a slew of projects in Neog.camp level zero curriculum. The program offers minimum hand holding and the available materials are enough to help whenever need arises.  Please visit my portfolio site for other such projects, I'll really appreciate and thank you for your time.